gather completion commands in single file
